+++ CHANGELOG.rst
@@ -1,6 +1,8 @@
 Version 0.0.10
 ==============
 
+* Changed the payment backend API to let get_finished_url and get_cancel_url
+  return strings instead of HttpResponse objects (this was confusing)
 * Tests for the shop are now runnable from any project
 
 Version 0.0.9